+++ internal/types/typed-script/index.ts
@@ -4,9 +4,11 @@ export {
   type CompileIssue,
   type CompileOptions,
   type CompileSuccess,
+  type CompileWarning,
 } from './compile';
 export {
   convertTsTypeToJsonSchema,
   type ConvertContext,
   type ConvertIssue,
+  type ConvertWarning,
 } from './ts-to-json-schema';